Enable CFF rich text support.
